refactor(auth): reorganize authentication context and update imports
- Moved authentication context to useAuthContext for better structure.
- Updated DialogLogin and DialogRegister components to reference the new context.
- Cleaned up AuthContext by removing unused code and improving readability.
