+++ front/src/contexts/useAuthContext.ts
@@ -10,7 +10,9 @@ export interface AuthContextType {
   refreshAuth: () => void;
 }
 
-export const AuthContext = createContext<AuthContextType | undefined>(undefined);
+export const AuthContext = createContext<AuthContextType | undefined>(
+  undefined
+);
 
 export const useAuth = () => {
   const context = useContext(AuthContext);
@@ -19,5 +21,3 @@ export const useAuth = () => {
   }
   return context;
 };
-
-
